+++ core/src/components/toast/toast.tsx
@@ -22,6 +22,8 @@ import { mdEnterAnimation } from './animations/md.enter';
 import { mdLeaveAnimation } from './animations/md.leave';
 import type { ToastAttributes, ToastPosition } from './toast-interface';
 
+// TODO(FW-2832): types
+
 /**
  * @virtualProp {"ios" | "md"} mode - The mode determines which platform styles to use.
  *
@@ -40,7 +42,7 @@ import type { ToastAttributes, ToastPosition } from './toast-interface';
   shadow: true,
 })
 export class Toast implements ComponentInterface, OverlayInterface {
-  private durationTimeout: any;
+  private durationTimeout?: ReturnType<typeof setTimeout>;
 
   presented = false;