+++ packages/react/src/components/IonAlert.tsx
@@ -1,4 +1,5 @@
-import { AlertOptions, alertController } from '@ionic/core/components';
+import type { AlertOptions } from '@ionic/core/components';
+import { alertController } from '@ionic/core/components';
 import { defineCustomElement } from '@ionic/core/components/ion-alert.js';
 
 import { createControllerComponent } from './createControllerComponent';