+++ packages/react/src/components/navigation/IonTabs.tsx
@@ -1,4 +1,4 @@
-import { JSX as LocalJSX } from '@ionic/core/components';
+import type { JSX as LocalJSX } from '@ionic/core/components';
 import React, { Fragment } from 'react';
 
 import { NavContext } from '../../contexts/NavContext';
@@ -7,13 +7,16 @@ import { HTMLElementSSR } from '../../utils/HTMLElementSSR';
 import { IonRouterOutlet } from '../IonRouterOutlet';
 
 import { IonTabBar } from './IonTabBar';
-import { IonTabsContext, IonTabsContextState } from './IonTabsContext';
+import type { IonTabsContextState } from './IonTabsContext';
+import { IonTabsContext } from './IonTabsContext';
 
 class IonTabsElement extends HTMLElementSSR {
   constructor() {
     super();
   }
 }
+
+// TODO(FW-2959): types
 
 if (typeof (window as any) !== 'undefined' && window.customElements) {
   const element = window.customElements.get('ion-tabs');
@@ -23,6 +26,7 @@ if (typeof (window as any) !== 'undefined' && window.customElements) {
 }
 
 declare global {
+  // eslint-disable-next-line @typescript-eslint/no-namespace
   namespace JSX {
     interface IntrinsicElements {
       'ion-tabs': any;
@@ -95,6 +99,7 @@ export const IonTabs = /*@__PURE__*/ (() =>
           : this.props.children;
 
       React.Children.forEach(children, (child: any) => {
+        // eslint-disable-next-line no-prototype-builtins
         if (child == null || typeof child !== 'object' || !child.hasOwnProperty('type')) {
           return;
         }
@@ -105,8 +110,8 @@ export const IonTabs = /*@__PURE__*/ (() =>
         }
 
         let childProps: any = {
-          ref: this.tabBarRef
-        }
+          ref: this.tabBarRef,
+        };
 
         /**
          * Only pass these props
@@ -118,15 +123,15 @@ export const IonTabs = /*@__PURE__*/ (() =>
         if (onIonTabsDidChange !== undefined) {
           childProps = {
             ...childProps,
-            onIonTabsDidChange
-          }
+            onIonTabsDidChange,
+          };
         }
 
         if (onIonTabsWillChange !== undefined) {
           childProps = {
             ...childProps,
-            onIonTabsWillChange
-          }
+            onIonTabsWillChange,
+          };
         }
 
         if (child.type === IonTabBar || child.type.isTabBar) {
@@ -149,11 +154,7 @@ export const IonTabs = /*@__PURE__*/ (() =>
       return (
         <IonTabsContext.Provider value={this.ionTabContextState}>
           {this.context.hasIonicRouter() ? (
-            <PageManager
-              className={className ? `${className}` : ''}
-              routeInfo={this.context.routeInfo}
-              {...props}
-            >
+            <PageManager className={className ? `${className}` : ''} routeInfo={this.context.routeInfo} {...props}>
               <ion-tabs className="ion-tabs" style={hostStyles}>
                 {tabBar.props.slot === 'top' ? tabBar : null}
                 <div style={tabsInner} className="tabs-inner">