+++ packages/react/src/components/routing-proxies.ts
@@ -8,32 +8,32 @@ import { IonItem as IonItemCmp } from '@ionic/core/components/ion-item.js';
 import { IonRouterLink as IonRouterLinkCmp } from '@ionic/core/components/ion-router-link.js';
 
 import { createRoutingComponent } from './createRoutingComponent';
-import { HrefProps } from './hrefprops';
+import type { HrefProps } from './hrefprops';
 
 export const IonRouterLink = /*@__PURE__*/ createRoutingComponent<
   HrefProps<JSX.IonRouterLink>,
   HTMLIonRouterLinkElement
 >('ion-router-link', IonRouterLinkCmp);
 
-export const IonButton = /*@__PURE__*/ createRoutingComponent<
-  HrefProps<JSX.IonButton>,
-  HTMLIonButtonElement
->('ion-button', IonButtonCmp);
+export const IonButton = /*@__PURE__*/ createRoutingComponent<HrefProps<JSX.IonButton>, HTMLIonButtonElement>(
+  'ion-button',
+  IonButtonCmp
+);
 
-export const IonCard = /*@__PURE__*/ createRoutingComponent<
-  HrefProps<JSX.IonCard>,
-  HTMLIonCardElement
->('ion-card', IonCardCmp);
+export const IonCard = /*@__PURE__*/ createRoutingComponent<HrefProps<JSX.IonCard>, HTMLIonCardElement>(
+  'ion-card',
+  IonCardCmp
+);
 
-export const IonFabButton = /*@__PURE__*/ createRoutingComponent<
-  HrefProps<JSX.IonFabButton>,
-  HTMLIonFabButtonElement
->('ion-fab-button', IonFabButtonCmp);
+export const IonFabButton = /*@__PURE__*/ createRoutingComponent<HrefProps<JSX.IonFabButton>, HTMLIonFabButtonElement>(
+  'ion-fab-button',
+  IonFabButtonCmp
+);
 
-export const IonItem = /*@__PURE__*/ createRoutingComponent<
-  HrefProps<JSX.IonItem>,
-  HTMLIonItemElement
->('ion-item', IonItemCmp);
+export const IonItem = /*@__PURE__*/ createRoutingComponent<HrefProps<JSX.IonItem>, HTMLIonItemElement>(
+  'ion-item',
+  IonItemCmp
+);
 
 export const IonItemOption = /*@__PURE__*/ createRoutingComponent<
   HrefProps<JSX.IonItemOption>,